Write unit tests for (new) util/ package.
This package contains two functions, one for writing text and one for binary-encoded decimal. The unit tests actually discovered some problems, mostly with WriteText. Firstly, the reserved space parameter did not consider the terminator. This was not discovered because the provided value was also not considering the terminator. I think this can be misleading, so the code was updated and godocs were added to clarify. The function will now return an error when insufficient space is provided.

Secondly, I wasn't checking the error upon Write() with the writer. I corrected it this but who knows where else I've made this mistake? I should introduce a linter so these cases are highlighted to me.